+++ src/ast/assignment-statement.ts
@@ -1,5 +1,5 @@
 import Context from '../semantics/context';
-import { IAstNode } from '../type-definitions/ast';
+import { IAstNode } from '../type-definitions/plainscript';
 
 export default class AssignmentStatement implements IAstNode<AssignmentStatement> {
   constructor(public targets: any[], public sources: any[]) { }
