+++ src/ast/function-declaration.ts
@@ -1,11 +1,11 @@
 import Context from '../semantics/context';
-import { Body, IAstNode, Parameter } from '../type-definitions/ast';
+import { Body, IAstNode } from '../type-definitions/plainscript';
 import FunctionObject from './function-object';
 // A function declaration binds a function object to a name.
 export default class FunctionDeclaration implements IAstNode<FunctionDeclaration> {
   public function: FunctionObject;
 
-  constructor(public id: string, params: Parameter, body: Body) {
+  constructor(public id: string, params: Entity, body: Body) {
     this.id = id;
     this.function = new FunctionObject(this.id, params, body);
   }
