+++ src/ast/if-statement.ts
@@ -1,5 +1,5 @@
 import Context from '../semantics/context';
-import { IAstNode, Statement } from '../type-definitions/ast';
+import { IAstNode, Statement } from '../type-definitions/plainscript';
 import BinaryExpression from './binary-expression';
 
 export default class IfStatement implements IAstNode<IfStatement> {
